requirements.txt
flask
python-dotenv
PyPDF2
sentence-transformers
faiss-cpu